refactor: move screenSizeHandler file
