+++ JS/controllers/handlers/screenSizeHandler.js
